pyproject.toml
[project]
name = "my-python-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.13",
    "Pillow>=11.0.0",
]